Convert SortableRuleList to TS
